+++ Docs/CHANGELOG.md
@@ -1,6 +1,12 @@
 # v0.0.2
-*What's new?*
+**What's new?**
 - Ask before downloading 
+- Maintain download history
 - Auto-ignore previously downloaded media
 
+# v0.0.3
+**What's new?**
+- Thread download process
+- Add media filepath in history
+- Zero stdout option while downloading